OutgoingClient: Close connection on unexpected stream to/from/id
If the client connects to a host serving another domain than expected,
usually the server will close the connection (as we are providing the
to attribute in our <stream> start). However, that must not happen.

To harden QXmpp a bit more, let's close the connection on any unexpected
event.

Also, checking the 'to' attribute prevents anything unexpected to
happen. The standard says the server must include it if we sent a from
attribute (which we do).