JingleIq/Rtp'Muting: Do not default init 'isMute' like 'creator' (breaking)
Otherwise this looks like the struct has default initializers for all
attributes and so one might forget to initialize creator (as happened in
the parsing code). Having a default to prefer one of the creator
enumerations (or having a default for isMute) doesn't really make sense.
People should just initialize both values.

Technically this is a breaking change, but I doubt anyone has actually
used this API yet (should be marked in changelog though).