OutgoingClient: Re-create socket on conn' error to avoid race conditions
Re-create socket if:
 1. TCP/TLS connection could be established to a host
 2. error occurs (on application level)
 3. next address is tried

I could reproduce one case where the socket needed 30s to connect to the
next host. In that case both the remote and we did close the connection
and that caused some race condition. Socket state and everything was
fine, but still some internal stuff was still waiting for a 30s timeout.

In general, reusing failed sockets is not a good idea and this could
improve stability.

However, on the other hand creating a new socket doesn't come without
costs and re-connecting to the same TLS host is sped up by reusing the
old socket.